Refactor and Implementation
Refactored the way manager classes use the Interfacer to query information and fully implemented the queries on the actual website. Also added "generateFakeData.py", which can be ran to fill a database with the fake user info. Users still have to manually edit the "config" in generateFakeData.py and __init__.py