Improvements and functionality
Added the ability for users to change their username and password in the settings page. Added ability for users to delete their account in the settings page. Refactored some methods of the UserManager class. Added the ability for users to add and remove friends from the users page when logged in.